requirements.txt
couchbase==4.1.2
python-dotenv==0.21.1
